<http://webkit.org/b/36878> REGRESSION: Trailing colon on hostnames (with no port specified) causes "Not allowed to use restricted network port"
Reviewed by Darin Adler.

WebCore:

* platform/KURL.cpp:
(WebCore::KURL::port): Explicitly handle the case of a colon being present in the URL after the host name but with
no port number before the path.  This is handled in the same manner as the colon and port being omitted completely.

LayoutTests:

Add explicit tests for handling of allowed ports.  We currently only test the ports that map to port 80: 0, 80, and "".

* platform/mac/security/block-test-expected.txt:
* security/block-test.html:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@56879 268f45cc-cd09-0410-ab3c-d52691b4dbfc